final review of grammar
Summary:
The final corrections of the grammar were made.
There were some minor issues, and some of them
helped to identify other errors in the syntax
diagrams. All of them were corrected.

An additional diagram for the methods
inside a specification was added. They
are methods without a body that serve as
the methods defined in a Java interface.

Issues:
The diagrams and grammar are not entirely
reviewed for ambiguity
and the grammar has not yet been compiled.