Fix compile errors
Summary:
The 'Type' token is a keyword in Python, therefore it cannot
be used in our language. The lexema remained, but the
rule changed to 'super_type'.

The 'Property' token is another keyword in Python 3.x.x
therfore its rule name was changed to 'field' as well
as its lexema.