Read arguments
Summary:
The parser is capable of creating the tables for
classes, specifications and enumerations with
their corresponding methods and arguments.

However, local variables declared within a method
are not yet supported.

Issues:
Even though it works correctly,
the key for the methods is only their name.
This will not enable us to define methods with
the same name, but different argument numbers
or types. This should be changed.